moving itParticlePositionReader, itkParticlePositionWriter, and itkParticleShapeStatistics to ParticleSystemUtils library
