+++ Libs/CMakeLists.txt
@@ -6,3 +6,4 @@ add_subdirectory(Transforms)
 add_subdirectory(SurfaceReconstruction)
 add_subdirectory(Utils)
 add_subdirectory(Vis)
+add_subdirectory(ParticleSystemUtils)
