+++ Post/source/ShapeWorksView2/CMakeLists.txt
@@ -180,8 +180,7 @@ ENDIF ( BUILD_SWV2_OSX_APP )
 SET( QT_LIBRARIES Qt5::Widgets Qt5::Core )
 
 
-TARGET_LINK_LIBRARIES( ShapeWorksView2 ${QT_LIBRARIES} ${ITK_LIBRARIES} tinyxml ${VTK_LIBRARIES})
-
+TARGET_LINK_LIBRARIES( ShapeWorksView2 ${QT_LIBRARIES} ${ITK_LIBRARIES} tinyxml ${VTK_LIBRARIES} ParticleSystemUtils)
 
 # INSTALLATION AND PACKAGING
 SET(plugin_dest_dir bin)
