+++ Run/source/ITKParticleSystem/CMakeLists.txt
@@ -5,12 +5,11 @@ FILE(GLOB SRCS "*.txx" "*.h" "*.cxx")
 ADD_LIBRARY (ITKParticleSystem STATIC
 ${SRCS}
 )
+target_link_libraries(ITKParticleSystem Utilities Procrustes ParticleSystemUtils)
 
 if(BUILD_MeshSupport OR BUILD_FeaMeshSupport)
   target_link_libraries(ITKParticleSystem fim_v4)
 endif(BUILD_MeshSupport OR BUILD_FeaMeshSupport)
-target_link_libraries(ITKParticleSystem Utilities Procrustes)
-
 
 # if(FIM_FOUND AND BUILD_MeshSupport)
   # target_link_libraries(ITKParticleSystem ${FIM_LIB})
